refactor(frontend): modularize remaining legacy page runtimes
