+++ js/legacy/game-view-init-loader.mjs
@@ -1,25 +1,12 @@
+import { bootGameView } from "../lib/game_view_runtime.js";
+
 const doc = globalThis.document;
-const El = globalThis.Element;
-
-import initGameView from "./modules/game-view-init.mjs";
-
-function boot(event) {
-    if (event?.type === "turbo:frame-load") {
-        const frame = event.target;
-        if (frame instanceof El) {
-            initGameView({ root: frame });
-            return;
-        }
-    }
-
-    initGameView({ root: doc });
-}
 
 if (doc.readyState === "loading") {
-    doc.addEventListener("DOMContentLoaded", boot);
+    doc.addEventListener("DOMContentLoaded", bootGameView);
 } else {
-    boot();
+    bootGameView();
 }
 
-doc.addEventListener("turbo:load", boot);
-doc.addEventListener("turbo:frame-load", boot);
+doc.addEventListener("turbo:load", bootGameView);
+doc.addEventListener("turbo:frame-load", bootGameView);
