+++ js/legacy/person-game-log-tabs-loader.mjs
@@ -1,28 +1,6 @@
+import { bootPersonGameLogTabs as boot } from "../lib/person_game_log_tabs_runtime.js";
+
 const doc = globalThis.document;
-const El = globalThis.Element;
-
-import initPersonGameLogTabs from "./modules/person-game-log-tabs.mjs";
-
-function getInit() {
-    const override = globalThis.__PERSON_GAME_LOG_TABS_INIT__;
-    if (typeof override === "function") {
-        return override;
-    }
-
-    return initPersonGameLogTabs;
-}
-
-function boot(event) {
-    if (event?.type === "turbo:frame-load") {
-        const frame = event.target;
-        if (frame instanceof El) {
-            getInit()({ root: frame });
-            return;
-        }
-    }
-
-    getInit()({ root: doc });
-}
 boot();
 
 doc.addEventListener("DOMContentLoaded", boot, { once: true });
